Reorganize install script
Complete overhaul oh the install script. It is now divided in small
modules for the installation step (script/modules) and for the
post-install processe (script/post). The install script also runs the
bootscrap script between the install and the post-install steps.

The post-install step adds first draft of ssh-keygen solution.